+++ conda/dependencies.go
@@ -21,7 +21,23 @@ type dependency struct {
 	Origin  string `yaml:"origin"  json:"channel"`
 }
 
+func (it *dependency) AsKey() string {
+	return fmt.Sprintf("%-50s  %-20s", it.Name, it.Origin)
+}
+
 type dependencies []*dependency
+
+func (it dependencies) sorted() dependencies {
+	sort.SliceStable(it, func(left, right int) bool {
+		lefty := strings.ToLower(it[left].Name)
+		righty := strings.ToLower(it[right].Name)
+		if lefty == righty {
+			return it[left].Origin < it[right].Origin
+		}
+		return lefty < righty
+	})
+	return it
+}
 
 func parseDependencies(origin string, output []byte) (dependencies, error) {
 	result := make(dependencies, 0, 100)
@@ -60,7 +76,7 @@ func fillDependencies(context, targetFolder string, seen map[string]string, coll
 	return collector, nil
 }
 
-func goldenMasterFilename(targetFolder string) string {
+func GoldenMasterFilename(targetFolder string) string {
 	return filepath.Join(targetFolder, "golden-ee.yaml")
 }
 
@@ -75,41 +91,98 @@ func goldenMaster(targetFolder string, pipUsed bool) (err error) {
 		collector, err = fillDependencies("pypi", targetFolder, seen, collector, "pip", "list", "--isolated", "--local", "--format", "json")
 		fail.On(err != nil, "Failed to list pip dependencies, reason: %v", err)
 	}
-	sort.SliceStable(collector, func(left, right int) bool {
-		return strings.ToLower(collector[left].Name) < strings.ToLower(collector[right].Name)
-	})
-	body, err := yaml.Marshal(collector)
+	body, err := yaml.Marshal(collector.sorted())
 	fail.On(err != nil, "Failed to make yaml, reason: %v", err)
-	goldenfile := goldenMasterFilename(targetFolder)
+	goldenfile := GoldenMasterFilename(targetFolder)
 	common.Debug("%sGolden EE file at: %v%s", pretty.Yellow, goldenfile, pretty.Reset)
 	return os.WriteFile(goldenfile, body, 0644)
 }
 
-func LoadWantedDependencies(filename string) (_ dependencies, err error) {
-	defer fail.Around(&err)
-
+func LoadWantedDependencies(filename string) dependencies {
 	body, err := os.ReadFile(filename)
-	fail.On(err != nil, "Failed to read dependencies from %q, reason: %v.", filename, err)
+	if err != nil {
+		return dependencies{}
+	}
 	result := make(dependencies, 0, 100)
 	err = yaml.Unmarshal(body, &result)
-	fail.On(err != nil, "Failed to parse dependencies from %q, reason: %v.", filename, err)
-	return result, nil
+	if err != nil {
+		return dependencies{}
+	}
+	return result.sorted()
 }
 
-func DumpEnvironmentDependencies(targetFolder string) (err error) {
+func SideBySideViewOfDependencies(goldenfile, wantedfile string) (err error) {
 	defer fail.Around(&err)
 
-	filename := goldenMasterFilename(targetFolder)
-	dependencyList, err := LoadWantedDependencies(filename)
-	fail.On(err != nil, "%v", err)
+	gold := LoadWantedDependencies(goldenfile)
+	want := LoadWantedDependencies(wantedfile)
 
+	if len(gold) == 0 && len(want) == 0 {
+		return fmt.Errorf("Running against old environment, and no dependencies.yaml.")
+	}
+
+	diffmap := make(map[string][2]int)
+	injectDiffmap(diffmap, want, 0)
+	injectDiffmap(diffmap, gold, 1)
+	keyset := make([]string, 0, len(diffmap))
+	for key, _ := range diffmap {
+		keyset = append(keyset, key)
+	}
+	sort.Strings(keyset)
+
+	common.WaitLogs()
+	hasgold := false
+	unknown := fmt.Sprintf("%sUnknown%s", pretty.Grey, pretty.Reset)
+	same := fmt.Sprintf("%sSame%s", pretty.Cyan, pretty.Reset)
+	drifted := fmt.Sprintf("%sDrifted%s", pretty.Yellow, pretty.Reset)
+	missing := fmt.Sprintf("%sN/A%s", pretty.Grey, pretty.Reset)
 	tabbed := tabwriter.NewWriter(os.Stderr, 2, 4, 2, ' ', 0)
-	tabbed.Write([]byte("No.\tPackage\tVersion\tOrigin\n"))
-	tabbed.Write([]byte("---\t-------\t-------\t------\n"))
-	for at, entry := range dependencyList {
-		data := fmt.Sprintf("%3d\t%s\t%s\t%s\n", at+1, entry.Name, entry.Version, entry.Origin)
+	tabbed.Write([]byte("Wanted\tVersion\tOrigin\t|\tNo.\t|\tAvailable\tVersion\tOrigin\t|\tStatus\n"))
+	tabbed.Write([]byte("------\t-------\t------\t+\t---\t+\t---------\t-------\t------\t+\t------\n"))
+	for at, key := range keyset {
+		//left, right, status := "n/a\tn/a\tn/a\t", "\tn/a\tn/a\tn/a", unknown
+		left, right, status := "-\t-\t-\t", "\t-\t-\t-", unknown
+		sides := diffmap[key]
+		if sides[0] < 0 || sides[1] < 0 {
+			status = missing
+		} else {
+			left, right := want[sides[0]], gold[sides[1]]
+			if left.Version == right.Version {
+				status = same
+			} else {
+				status = drifted
+			}
+		}
+		if sides[0] > -1 {
+			entry := want[sides[0]]
+			left = fmt.Sprintf("%s\t%s\t%s\t", entry.Name, entry.Version, entry.Origin)
+		}
+		if sides[1] > -1 {
+			entry := gold[sides[1]]
+			right = fmt.Sprintf("\t%s\t%s\t%s", entry.Name, entry.Version, entry.Origin)
+			hasgold = true
+		}
+		data := fmt.Sprintf("%s|\t%3d\t|%s\t|\t%s\n", left, at+1, right, status)
 		tabbed.Write([]byte(data))
 	}
+	tabbed.Write([]byte("------\t-------\t------\t+\t---\t+\t---------\t-------\t------\t+\t------\n"))
+	tabbed.Write([]byte("Wanted\tVersion\tOrigin\t|\tNo.\t|\tAvailable\tVersion\tOrigin\t|\tStatus\n"))
+	tabbed.Write([]byte("\n"))
 	tabbed.Flush()
+	if !hasgold {
+		return fmt.Errorf("Running against old environment, which does not have 'golden-ee.yaml' file.")
+	}
 	return nil
+}
+
+func injectDiffmap(diffmap map[string][2]int, deps dependencies, side int) {
+	for at, entry := range deps {
+		key := entry.AsKey()
+		found, ok := diffmap[key]
+		if !ok {
+			found = [2]int{-1, -1}
+		}
+		found[side] = at
+		diffmap[key] = found
+	}
 }